requirements.txt
ImageHash==4.3.1
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
psutil==6.0.0
pytesseract==0.3.13
PyWavelets==1.7.0
scipy==1.14.1
